requirements.txt
aiogram==2.25.1
aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.30.1
pytz==2024.1
